Convert MainActivity.kt to Composable UI This marks the initial fully-composable mobile module. There's still a lot of styles to touch up, and it might pay to create some more UI tests
