+++ mobile/src/main/java/com/boswelja/devicemanager/aboutapp/ui/AppInfoViewModel.kt
@@ -5,7 +5,7 @@
  * This file, and any part of the Wearable Extensions app/s cannot be copied and/or distributed
  * without permission from Jack Boswell (boswelja) <boswela@outlook.com>
  */
-package com.boswelja.devicemanager.appinfo.ui
+package com.boswelja.devicemanager.aboutapp.ui
 
 import android.app.Application
 import androidx.browser.customtabs.CustomTabsIntent
