+++ mobile/src/main/java/com/boswelja/devicemanager/batterysync/ui/BatterySyncSettingsWidgetFragment.kt
@@ -1,9 +1,5 @@
 package com.boswelja.devicemanager.batterysync.ui
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.aspectRatio
@@ -17,35 +13,15 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.boswelja.devicemanager.R
 import com.boswelja.devicemanager.batterysync.database.WatchBatteryStats
 import com.boswelja.devicemanager.common.BatteryIcon
-import com.boswelja.devicemanager.common.ui.AppTheme
 import java.util.concurrent.TimeUnit
-
-class BatterySyncPreferenceWidgetFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        return ComposeView(requireContext()).apply {
-            setContent {
-                AppTheme {
-                    BatterySyncSettingsWidget()
-                }
-            }
-        }
-    }
-}
 
 @Composable
 fun BatterySyncSettingsWidget() {
