--- mobile/src/main/res/layout/common_toolbar.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<com.google.android.material.appbar.AppBarLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/toolbar_layout"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    app:liftOnScroll="true">
-
-    <com.google.android.material.appbar.MaterialToolbar
-        android:id="@+id/toolbar"
-        style="@style/Widget.MaterialComponents.Toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:animateLayoutChanges="true"
-        android:clipToPadding="false"
-        android:minHeight="?attr/actionBarSize" />
-</com.google.android.material.appbar.AppBarLayout>
