--- mobile/src/main/res/layout/watch_selector_item.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.appcompat.widget.AppCompatTextView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/title"
-    style="@style/TextAppearance.MaterialComponents.Body1"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:paddingStart="@dimen/watch_selector_horizontal_margin"
-    android:paddingTop="@dimen/watch_selector_vertical_margin"
-    android:paddingEnd="@dimen/watch_selector_horizontal_margin"
-    android:paddingBottom="@dimen/watch_selector_vertical_margin"
-    android:textColor="?android:textColorPrimary" />
-
