--- mobile/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:id="@id/dashboardFragment"
-        android:icon="@drawable/nav_ic_dashboard"
-        android:title="@string/bottom_nav_dashboard_label" />
-
-    <item
-        android:id="@id/messagesFragment"
-        android:icon="@drawable/nav_ic_message"
-        android:title="@string/nav_messages_label" />
-
-    <item
-        android:id="@id/appSettingsFragment"
-        android:icon="@drawable/nav_ic_app_settings"
-        android:title="@string/bottom_nav_app_settings_label" />
-
-    <item
-        android:id="@id/appInfoFragment"
-        android:icon="@drawable/nav_ic_about"
-        android:title="@string/bottom_nav_about_label" />
-
-</menu>
