--- mobile/src/main/res/navigation/main_graph.xml
@@ -1,115 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/main_navigation"
-    app:startDestination="@id/dashboardFragment">
-
-    <action
-        android:id="@+id/action_global_dashboardFragment"
-        app:destination="@id/dashboardFragment" />
-
-    <action
-        android:id="@+id/action_global_messagesFragment"
-        app:destination="@id/messagesFragment" />
-
-    <action
-        android:id="@+id/action_global_appSettingsFragment"
-        app:destination="@id/appSettingsFragment" />
-
-    <action
-        android:id="@+id/action_global_appInfoFragment"
-        app:destination="@id/appInfoFragment" />
-
-    <fragment
-        android:id="@+id/dashboardFragment"
-        android:name="com.boswelja.devicemanager.dashboard.ui.DashboardFragment">
-
-        <action
-            android:id="@+id/to_batterySyncActivity"
-            app:destination="@id/batterySyncPreferenceActivity" />
-
-        <action
-            android:id="@+id/to_dndSyncActivity"
-            app:destination="@id/dnDSyncPreferenceActivity" />
-
-        <action
-            android:id="@+id/to_phoneLockingActivity"
-            app:destination="@id/phoneLockingPreferenceActivity" />
-
-        <action
-            android:id="@+id/to_appManagerActivity"
-            app:destination="@id/appManagerActivity" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/messagesFragment"
-        android:name="com.boswelja.devicemanager.messages.ui.MessagesFragment">
-
-        <action
-            android:id="@+id/to_messageHistoryActivity"
-            app:destination="@id/messageHistoryActivity" />
-
-    </fragment>
-
-    <fragment
-        android:id="@+id/appSettingsFragment"
-        android:name="com.boswelja.devicemanager.appsettings.ui.AppSettingsFragment">
-
-        <action
-            android:id="@+id/to_widgetSettingsActivity"
-            app:destination="@id/widgetSettingsActivity" />
-
-        <action
-            android:id="@+id/to_watchManagerActivity"
-            app:destination="@id/watchManagerActivity" />
-
-        <action
-            android:id="@+id/to_manageSpaceActivity"
-            app:destination="@id/manageSpaceActivity" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/appInfoFragment"
-        android:name="com.boswelja.devicemanager.appinfo.ui.AppInfoFragment">
-
-        <action
-            android:id="@+id/to_donateActivity"
-            app:destination="@id/donateActivity" />
-    </fragment>
-
-    <activity
-        android:id="@+id/donateActivity"
-        android:name="com.boswelja.devicemanager.donate.ui.DonateActivity" />
-
-    <activity
-        android:id="@+id/batterySyncPreferenceActivity"
-        android:name="com.boswelja.devicemanager.batterysync.ui.BatterySyncSettingsActivity" />
-
-    <activity
-        android:id="@+id/appManagerActivity"
-        android:name="com.boswelja.devicemanager.appmanager.ui.AppManagerActivity" />
-
-    <activity
-        android:id="@+id/messageHistoryActivity"
-        android:name="com.boswelja.devicemanager.messages.ui.MessageHistoryActivity" />
-
-    <activity
-        android:id="@+id/dnDSyncPreferenceActivity"
-        android:name="com.boswelja.devicemanager.dndsync.ui.DnDSyncPreferenceActivity" />
-
-    <activity
-        android:id="@+id/phoneLockingPreferenceActivity"
-        android:name="com.boswelja.devicemanager.phonelocking.ui.PhoneLockingSettingsActivity" />
-
-    <activity
-        android:id="@+id/widgetSettingsActivity"
-        android:name="com.boswelja.devicemanager.widget.ui.WidgetSettingsActivity" />
-
-    <activity
-        android:id="@+id/watchManagerActivity"
-        android:name="com.boswelja.devicemanager.watchmanager.ui.WatchManagerActivity" />
-
-    <activity
-        android:id="@+id/manageSpaceActivity"
-        android:name="com.boswelja.devicemanager.managespace.ui.ManageSpaceActivity" />
-</navigation>
