+++ mobile/src/main/res/values/dimens.xml
@@ -3,11 +3,5 @@
 
     <dimen name="corner_radius">8dp</dimen>
 
-    <dimen name="watch_selector_horizontal_margin">16dp</dimen>
-    <dimen name="watch_selector_vertical_margin">8dp</dimen>
-
-    <dimen name="dashboard_card_spacing">4dp</dimen>
-    <dimen name="dashboard_content_padding">16dp</dimen>
-
     <dimen name="rotation_offset">0</dimen>
 </resources>
