+++ mobile/src/main/res/values/styles.xml
@@ -1,63 +1,5 @@
 <resources>
 
-    <style name="Theme.Base" parent="Theme.MaterialComponents.DayNight.NoActionBar">
-        <item name="colorPrimary">@color/colorPrimary</item>
-        <item name="colorPrimaryDark">@color/statusBarColor</item>
-        <item name="colorSecondary">@color/colorPrimary</item>
-        <item name="colorOnSecondary">@color/primaryTextColor</item>
-        <item name="colorAccent">@color/colorPrimary</item>
-        <item name="android:windowBackground">@color/windowBackground</item>
-        <item name="alertDialogTheme">@style/Theme.App.Dialog</item>
-        <item name="toolbarStyle">@style/Theme.App.Toolbar</item>
-        <item name="bottomSheetDialogTheme">@style/Theme.App.BottomSheetDialog</item>
-        <item name="materialCardViewStyle">@style/Widget.App.CardView</item>
-        <item name="shapeAppearanceSmallComponent">@style/ShapeAppearance.App.SmallComponent</item>
-        <item name="shapeAppearanceMediumComponent">@style/ShapeAppearance.App.MediumComponent
-        </item>
-        <item name="shapeAppearanceLargeComponent">@style/ShapeAppearance.App.LargeComponent</item>
-    </style>
-
-    <style name="Theme.App" parent="Theme.Base" />
-
-    <style name="Theme.App.Toolbar" parent="Widget.MaterialComponents.Toolbar">
-        <item name="android:background">@color/windowBackground</item>
-    </style>
-
-    <style name="Theme.App.Dialog" parent="Theme.MaterialComponents.DayNight.Dialog">
-        <item name="android:windowBackground">@drawable/dialog_background</item>
-        <item name="colorAccent">@color/colorPrimary</item>
-        <item name="colorPrimary">@color/colorPrimary</item>
-    </style>
-
-    <style name="Theme.App.BottomSheetDialog" parent="Theme.MaterialComponents.DayNight.BottomSheetDialog">
-        <item name="bottomSheetStyle">@style/Widget.App.BottomSheet</item>
-    </style>
-
-    <style name="ShapeAppearance.App.SmallComponent" parent="ShapeAppearance.MaterialComponents.SmallComponent">
-        <item name="cornerFamily">rounded</item>
-        <item name="cornerSize">@dimen/corner_radius</item>
-    </style>
-
-    <style name="ShapeAppearance.App.MediumComponent" parent="ShapeAppearance.MaterialComponents.MediumComponent">
-        <item name="cornerFamily">rounded</item>
-        <item name="cornerSize">@dimen/corner_radius</item>
-    </style>
-
-    <style name="ShapeAppearance.App.LargeComponent" parent="ShapeAppearance.MaterialComponents.LargeComponent">
-        <item name="cornerFamily">rounded</item>
-        <item name="cornerSize">@dimen/corner_radius</item>
-        <item name="android:backgroundTint">@color/dialogBackground</item>
-    </style>
-
-    <style name="Widget.App.BottomSheet" parent="Widget.MaterialComponents.BottomSheet.Modal">
-        <item name="backgroundTint">@color/dialogBackground</item>
-    </style>
-
-    <style name="Widget.App.CardView" parent="Widget.MaterialComponents.CardView">
-        <item name="cardElevation">0dp</item>
-        <item name="strokeColor">@color/dividerColor</item>
-        <item name="strokeWidth">1dp</item>
-        <item name="cardBackgroundColor">@color/windowBackground</item>
-    </style>
+    <style name="Theme.App" parent="Theme.MaterialComponents.DayNight.NoActionBar" />
 
 </resources>
